Testing and Bug Fixes
Fixed issues with gaussian and polynomial fits, and added preliminary test
file. Also rewrote CSV functions and made general improvements to
parameter estimation.

Issues:
- The test class needs to be expanded to incorporate physical datasets
rather than testing fits to known functions, and to improve coverage.
The functions all appear to work at present, but there are likely still
bugs to iron out.
- The curve fit algorithm fails to estimate covariance for polynomial
fits when some terms have coefficients of 0. This is unlikely to be an
issue for data with noise, however.
- If the user passes in a map, filter or similar object as a sequence of
data, _general_fit throws a ValueError that is not indicative of the
cause. Should be relatively easy to correct.