Added function for generating LaTeX tables
Added the write_to_Tex function, which writes the input data to a file in LaTeX tabular format. The output could use some formatting, but it should be functional
at the moment.

Also:
Modified the CSV functions so they would be suitable for use with the
Tex function.
Improved keyword argument handling in the draw_plot and save_figure
functions.